Moving make_stat_first_coordinate to cube_manipulation.py
