+++ lib/improver/tests/spotdata/spotdata/test_extract_data.py
@@ -32,7 +32,6 @@
 
 import numpy as np
 import unittest
-import warnings
 
 from collections import OrderedDict
 from datetime import datetime as dt
@@ -50,8 +49,6 @@ from iris.tests import IrisTest
 from iris.time import PartialDateTime
 from iris.exceptions import CoordinateNotFoundError
 
-from improver.tests.nbhood.nbhood.test_NeighbourhoodProcessing import (
-    set_up_cube)
 from improver.spotdata.extract_data import ExtractData as Plugin
 
 
@@ -331,65 +328,6 @@ class Test__build_coordinate(Test_setup):
         self.assertEqual(result.name(), 'latitude')
         self.assertIsInstance(result, DimCoord)
         self.assertEqual(result.units, 'degrees')
-
-
-class Test_make_stat_coordinate_first(Test_setup):
-    """Test the function to reorder a cube to ensure the statistical coordinate
-    is the first dimension."""
-
-    def test_cube_reorder_realization(self):
-        """Test reordering a cube with a realization coordinate to make it come
-        first."""
-
-        plugin = Plugin().make_stat_coordinate_first
-        cube = set_up_cube(
-            zero_point_indices=(
-                (0, 0, 2, 2), (1, 0, 3, 3), (0, 1, 0, 0), (1, 1, 2, 1)),
-            num_time_points=2, num_grid_points=5, num_realization_points=2)
-        incorrect_cube = cube.copy()
-        incorrect_cube.transpose([1, 2, 0, 3])
-        result = plugin(incorrect_cube)
-        self.assertEqual(cube.coords()[0].name(), result.coords()[0].name())
-        self.assertEqual(cube.coords(), result.coords())
-
-    def test_cube_reorder_percentile(self):
-        """Test reordering a cube with a percentile coordinate to make it come
-        first."""
-
-        plugin = Plugin().make_stat_coordinate_first
-        cube = set_up_cube(
-            zero_point_indices=(
-                (0, 0, 2, 2), (1, 0, 3, 3), (0, 1, 0, 0), (1, 1, 2, 1)),
-            num_time_points=2, num_grid_points=5, num_realization_points=2)
-        cube.coords()[0].rename('percentile_over_time')
-        incorrect_cube = cube.copy()
-        incorrect_cube.transpose([1, 2, 0, 3])
-        result = plugin(incorrect_cube)
-        self.assertEqual(cube.coords()[0].name(), result.coords()[0].name())
-        self.assertEqual(cube.coords(), result.coords())
-
-    def test_cube_reorder_percentile_and_realization(self):
-        """Test reordering a cube with a percentile and a realization
-        coordinate. Should produce a warning and promote the first statistical
-        coordinate that is found."""
-
-        plugin = Plugin().make_stat_coordinate_first
-        cube = set_up_cube(
-            zero_point_indices=(
-                (0, 0, 2, 2), (1, 0, 3, 3), (0, 1, 0, 0), (1, 1, 2, 1)),
-            num_time_points=2, num_grid_points=5, num_realization_points=2)
-        realization = cube.coords()[0][0].copy()
-        cube.coords()[0].rename('percentile_over_time')
-        cube.add_aux_coord(realization)
-        cube = iris.util.new_axis(cube, 'realization')
-        incorrect_cube = cube.copy()
-        incorrect_cube.transpose([2, 1, 0, 3, 4])
-        with warnings.catch_warnings(record=True) as w_messages:
-            result = plugin(incorrect_cube)
-            assert len(w_messages) == 1
-            assert issubclass(w_messages[0].category, UserWarning)
-            assert "More than one statistical" in str(w_messages[0])
-        self.assertEqual(cube.coords()[0].name(), result.coords()[0].name())
 
 
 class Test_make_cube(Test_setup):
